fix: add /api/auth/session endpoint for reliable user provisioning
Replace scattered server actions (ensureUser, verifyUser, seedDefaults)
with a single GET /api/auth/session endpoint that handles the full
auth flow: session validation, user upsert, and default seeding.

The dashboard layout now calls this endpoint on mount — if it returns
401, the user is signed out. The login page calls it after auth to
provision the user before redirecting.

This fixes the "User not found" errors that occurred when the DB was
reset but the auth session was still valid — the old approach silently
swallowed provisioning errors and rendered a broken dashboard.

Also consolidates duplicate generateApiKey/generateAccessToken functions
into single exported definitions in their respective service files.
