+++ apps/web/src/app/(dashboard)/layout.tsx
@@ -1,21 +1,23 @@
 "use client";
 
-import { useEffect } from "react";
+import { useEffect, useState, useRef, useCallback } from "react";
 import { useRouter } from "next/navigation";
 import { ScrollArea } from "@onecli/ui/components/scroll-area";
 import { SidebarInset, SidebarProvider } from "@onecli/ui/components/sidebar";
 import { DashboardSidebar } from "@dashboard/dashboard-sidebar";
 import { DashboardHeader } from "@dashboard/dashboard-header";
 import { useAuth } from "@/providers/auth-provider";
-import { seedDefaults } from "@/lib/actions/auth";
 
 export default function DashboardLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const { isAuthenticated, isLoading } = useAuth();
+  const { isAuthenticated, isLoading, signOut } = useAuth();
   const router = useRouter();
+  const [ready, setReady] = useState(false);
+  const signOutRef = useRef(signOut);
+  signOutRef.current = signOut;
 
   useEffect(() => {
     if (!isLoading && !isAuthenticated) {
@@ -23,13 +25,26 @@ export default function DashboardLayout({
     }
   }, [isLoading, isAuthenticated, router]);
 
+  const initSession = useCallback(async () => {
+    try {
+      const res = await fetch("/api/auth/session");
+      if (res.ok) {
+        setReady(true);
+      } else {
+        signOutRef.current();
+      }
+    } catch {
+      signOutRef.current();
+    }
+  }, []);
+
   useEffect(() => {
     if (isAuthenticated) {
-      seedDefaults();
+      initSession();
     }
-  }, [isAuthenticated]);
+  }, [isAuthenticated, initSession]);
 
-  if (isLoading) {
+  if (isLoading || (isAuthenticated && !ready)) {
     return (
       <div className="flex h-svh items-center justify-center">
         <div className="text-brand h-8 w-8 animate-spin rounded-full border-2 border-current border-t-transparent" />
